+++ docs/conf.py
@@ -19,7 +19,7 @@ project = "COMPAS"
 copyright = "Block Research Group - ETH Zurich"
 author = "Tom Van Mele"
 
-release = "0.18.1"
+release = "0.19.0"
 version = ".".join(release.split(".")[0:2])
 
 master_doc = "index"
@@ -28,7 +28,7 @@ source_suffix = {
     ".md": "markdown",
 }
 templates_path = ["_templates", ]
-exclude_patterns = ["_build", "**.ipynb_checkpoints", "_notebooks"]
+exclude_patterns = ["_build", "**.ipynb_checkpoints", "_notebooks", "**/__temp"]
 
 pygments_style = "sphinx"
 show_authors = True
@@ -48,6 +48,8 @@ extensions = [
     "sphinx.ext.viewcode",
     "sphinx.ext.githubpages",
     "sphinx.ext.coverage",
+    "sphinx.ext.inheritance_diagram",
+    # "sphinx.ext.graphviz",
     "matplotlib.sphinxext.plot_directive",
     "m2r",
     "nbsphinx",
@@ -92,6 +94,11 @@ autosummary_mock_imports = [
     "bmesh",
     "mathutils"
 ]
+
+# graph options
+
+inheritance_graph_attrs = dict(rankdir="TB", size='"16.0, 16.0"', ratio="auto")
+inheritance_node_attrs = dict(fontsize=14)
 
 # napoleon options
 
@@ -251,20 +258,13 @@ intersphinx_mapping = {
 
 # -- Options for HTML output ----------------------------------------------
 
-package_docs_root = "/compas/"
-
-with open(os.path.join(os.path.dirname(__file__), "docversions.txt"), "r") as f:
-    version_names = [version.strip() for version in f.readlines()]
-    package_docs_versions = [(version, "{}{}".format(package_docs_root, version))
-                             for version in version_names if version]
-
 html_theme = "compas"
 html_theme_path = sphinx_compas_theme.get_html_theme_path()
 html_theme_options = {
     "navbar_active": "compas",
     "package_version": release,
     "package_docs": "https://compas.dev/compas/",
-    "package_old_versions_txt": "https://compas.dev/compas/docversions.txt"
+    "package_old_versions_txt": "https://compas.dev/compas/doc_versions.txt"
 }
 html_context = {}
 html_static_path = []