+++ src/compas_ghpython/artists/frameartist.py
@@ -2,8 +2,6 @@ from __future__ import print_function
 from __future__ import absolute_import
 from __future__ import division
 
-import itertools
-
 import compas_ghpython
 from compas_ghpython.artists._primitiveartist import PrimitiveArtist
 
@@ -60,31 +58,6 @@ class FrameArtist(PrimitiveArtist):
         point, lines = self._get_args(self.primitive, self.scale, self.color_origin, self.color_xaxis, self.color_yaxis, self.color_zaxis)
         geometry = [None, None]
         geometry[0] = compas_ghpython.draw_points([point])
-        geometry[1] = compas_ghpython.draw_lines(lines)
-        return geometry
-
-    @staticmethod
-    def draw_collection(collection):
-        """Draw the collections of frames.
-
-        Parameters
-        ----------
-        collection : list of compas.geometry.Frame
-            A collection of ``Frame`` objects.
-
-        Returns
-        -------
-        geometry : list
-
-            * geometry[0] : list of :class:`Rhino.Geometry.Point`
-            * geometry[1] : list of :class:`Rhino.Geometry.Line`
-
-        """
-        args = [FrameArtist._get_args(primitive) for primitive in collection]
-        points, lines = zip(*args)
-        lines = itertools.chain(*lines)
-        geometry = [None, None]
-        geometry[0] = compas_ghpython.draw_points(points)
         geometry[1] = compas_ghpython.draw_lines(lines)
         return geometry