+++ src/compas_plotters/__init__.py
@@ -19,7 +19,7 @@ Classes
 
 """
 
-__version__ = '0.18.1'
+__version__ = '0.19.0'
 
 from .core import *  # noqa: F401 F403
 from .artists import *  # noqa: F401 F403