+++ src/compas_rhino/__init__.py
@@ -29,7 +29,7 @@ if compas.RHINO:
     from .utilities import *  # noqa: F401 F403
 
 
-__version__ = '0.18.1'
+__version__ = '0.19.0'
 
 
 PURGE_ON_DELETE = True